+++ example/app/pages/frame/pages/onlineMap/pages/gaode/gaode.component.ts
@@ -3,7 +3,7 @@ import ViewerOptions = Cesium.ViewerOptions;
 import Viewer = Cesium.Viewer;
 import Scene = Cesium.Scene;
 import Globe = Cesium.Globe;
-import { GaodeImageryProvider, GaodeMapsStyle } from '../../../../../../../../src';
+import { GaodeMapsImageryProvider, GaodeMapsStyle } from '../../../../../../../../src';
 import { SelectItem } from 'primeng/primeng';
 
 @Component({
@@ -16,17 +16,17 @@ export class GaodeComponent implements OnInit {
 	scene: Scene;
 	globe: Globe;
 	servers: SelectItem[];
-	selectedServer: GaodeImageryProvider;
+	selectedServer: GaodeMapsImageryProvider;
 
 	constructor() {
 		this.viewerOptions = {
-			imageryProvider: new GaodeImageryProvider(GaodeMapsStyle.IMG)
+			imageryProvider: new GaodeMapsImageryProvider(GaodeMapsStyle.IMG)
 		};
 		this.servers = [
 			{label: '--选择服务类型--', value: null},
-			{label: '矢量地图服务', value: new GaodeImageryProvider(GaodeMapsStyle.VEC)},
-			{label: '影像地图服务', value: new GaodeImageryProvider(GaodeMapsStyle.IMG)},
-			{label: '影像标注服务', value: new GaodeImageryProvider(GaodeMapsStyle.CIA)}
+			{label: '矢量地图服务', value: new GaodeMapsImageryProvider(GaodeMapsStyle.VEC)},
+			{label: '影像地图服务', value: new GaodeMapsImageryProvider(GaodeMapsStyle.IMG)},
+			{label: '影像标注服务', value: new GaodeMapsImageryProvider(GaodeMapsStyle.CIA)}
 		];
 	}
 
@@ -41,11 +41,11 @@ export class GaodeComponent implements OnInit {
 
 	serverChange($event) {
 		this.viewer.imageryLayers.removeAll();
-		let gaodeImageryProvider: GaodeImageryProvider;
+		let gaodeImageryProvider: GaodeMapsImageryProvider;
 		if ($event.value) {
 			gaodeImageryProvider = $event.value;
 		} else {
-			gaodeImageryProvider = new GaodeImageryProvider(GaodeMapsStyle.IMG);
+			gaodeImageryProvider = new GaodeMapsImageryProvider(GaodeMapsStyle.IMG);
 		}
 		this.viewer.imageryLayers.addImageryProvider(gaodeImageryProvider);
 	}